refactor: extract transcript views from AgentWorkspace
